requirements.txt
flask
flask-cors
requests
python-dotenv
nanoid
python-telegram-bot
urllib3
